+++ packages/certs-v5/lib/endpoints.js
@@ -1,34 +1,15 @@
 'use strict'
 
-const { checkPrivateSniFeature } = require('./features.js')
-
-function sslCertsPromise (app, heroku) {
-  return heroku.request({
-    path: `/apps/${app}/ssl-endpoints`,
-    headers: {'Accept': 'application/vnd.heroku+json; version=3.ssl_cert'}
-  }).then(function (data) {
-    return data
-  })
-}
-
 function sniCertsPromise (app, heroku) {
-  return heroku.request({path: `/apps/${app}/sni-endpoints`}).catch(function (err) {
-    if (err.statusCode === 422 && err.body && err.body.id === 'space_app_not_supported') {
-      return []
-    }
-    throw err
-  }).then(function (data) {
+  return heroku.request({path: `/apps/${app}/sni-endpoints`}).then(function (data) {
     return data
   })
 }
 
 function meta (app, t, name) {
-  var path, variant
+  var path
   if (t === 'sni') {
     path = `/apps/${app}/sni-endpoints`
-  } else if (t === 'ssl') {
-    path = `/apps/${app}/ssl-endpoints`
-    variant = 'ssl_cert'
   } else {
     throw Error('Unknown type ' + t)
   }
@@ -36,57 +17,23 @@ function meta (app, t, name) {
     path = `${path}/${name}`
   }
 
-  if (variant) {
-    return {path, variant, flag: t}
-  } else {
-    return {path, flag: t}
-  }
+  return {path, flag: t}
 }
 
-function tagAndSort (app, allCerts) {
-  allCerts.sni_certs.forEach(function (cert) {
+function tagAndSort (app, sniCerts) {
+  sniCerts.forEach(function (cert) {
     cert._meta = meta(app, 'sni', cert.name)
   })
 
-  allCerts.ssl_certs.forEach(function (cert) {
-    cert._meta = meta(app, 'ssl', cert.name)
-  })
-
-  return allCerts.ssl_certs.concat(allCerts.sni_certs).sort(function (a, b) {
+  return sniCerts.sort(function (a, b) {
     return a.name < b.name
   })
 }
 
 async function all(appName, heroku) {
-  const featureList = await heroku.get(`/apps/${appName}/features`)
-  const privateSniFeatureEnabled = checkPrivateSniFeature(featureList)
+  let sniCerts = await sniCertsPromise(appName, heroku)
 
-  let [ssl_certs, sni_certs] = await Promise.all([
-    // use SNI endpoints only
-    privateSniFeatureEnabled ? [] : sslCertsPromise(appName, heroku),
-    sniCertsPromise(appName, heroku)
-  ])
-
-  let allCerts = {
-    ssl_certs,
-    sni_certs
-  }
-
-  return tagAndSort(appName, allCerts)
-}
-
-async function hasAddon(app, heroku) {
-  return await heroku.request({
-    path: `/apps/${app}/addons/ssl%3Aendpoint`
-  }).then(function () {
-    return true
-  }).catch(function (err) {
-    if (err.statusCode === 404 && err.body && err.body.id === 'not_found' && err.body.resource === 'addon') {
-      return false
-    } else {
-      throw err
-    }
-  });
+  return tagAndSort(appName, sniCerts)
 }
 
 async function hasSpace(app, heroku) {
@@ -99,7 +46,6 @@ async function hasSpace(app, heroku) {
 
 module.exports = {
   hasSpace,
-  hasAddon,
   meta,
   all
 }